requirements.txt
openai
python-dotenv
supabase
fastapi 
PyPDF2